Move Canonical class to Routing folder/namespace.
