+++ src/EventListener/SnippetListener.php
@@ -6,9 +6,9 @@ use Bolt\Asset\QueueInterface;
 use Bolt\Asset\Snippet\Queue as SnippetQueue;
 use Bolt\Asset\Snippet\Snippet;
 use Bolt\Asset\Target;
-use Bolt\Canonical;
 use Bolt\Config;
 use Bolt\Controller\Zone;
+use Bolt\Routing\Canonical;
 use Symfony\Component\Asset\Packages;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
