+++ src/Twig/Runtime/RoutingRuntime.php
@@ -2,8 +2,8 @@
 
 namespace Bolt\Twig\Runtime;
 
-use Bolt\Canonical;
 use Bolt\Library;
+use Bolt\Routing\Canonical;
 use Symfony\Component\HttpFoundation\RequestStack;
 
 /**
